JSON powered repositories list with links
The site was pretty bare and the list was no even linked.
The list is now powered by JSON and a vue component, meaning
we can add automation, and pretty display of these things as
we continue to move forward.
Repositories also now exit on their own page rather than the
home page.