Extract more things from main journey docs..
Before this there was still lots of duplication around what happened.
Infact the same sentence was in about 3 different places on multiple occasions.

Hopefully this is now a little easier to digest.
Also the separate hike and trailblaze pages were merged
into a generic journey one as there are not so many differences
it seems.